feat : 연차(반차 포함) 갯수 증감 및 휴가 일정 변경

+++ src/main/java/com/example/demo/employee/dto/response/AnnualLeaveResponseDto.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class AnnualLeaveResponseDto {
     private String employeeName;
-    private Integer annualLeave;
+    private Double annualLeave;
 }
